requirements.txt
absl-py==2.2.2
certifi==2025.1.31
charset-normalizer==3.4.1
filelock==3.18.0
fsspec==2025.3.2
grpcio==1.71.0
idna==3.10
importlib_metadata==8.6.1
imutils==0.5.4
Jinja2==3.1.6
kagglehub==0.3.12
Markdown==3.8
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
protobuf==6.30.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==1.26.6
Werkzeug==3.1.3
zipp==3.21.0
